Reverted serialize tables

+++ src/Models.fs
@@ -50,7 +50,7 @@ type UserUpd =
     {
         first_name : string
         last_name: string
-        birth_date: Nullable<int64>
+        birth_date: Nullable<int32>
         gender: Nullable<char>
         email: string
     }
@@ -70,7 +70,7 @@ type User() =
         [<DefaultValue>]val mutable id: int32
         [<DefaultValue>]val mutable first_name : byte[]
         [<DefaultValue>]val mutable last_name: byte[]
-        [<DefaultValue>]val mutable birth_date: int64
+        [<DefaultValue>]val mutable birth_date: int32
         [<DefaultValue>]val mutable gender: char
         [<DefaultValue>]val mutable email: byte[]
 
@@ -78,7 +78,7 @@ type UserOld() =
         [<DefaultValue>]val mutable id: int32
         [<DefaultValue>]val mutable first_name : string
         [<DefaultValue>]val mutable last_name: string
-        [<DefaultValue>]val mutable birth_date: int64
+        [<DefaultValue>]val mutable birth_date: int32
         [<DefaultValue>]val mutable gender: char
         [<DefaultValue>]val mutable email: string
